+++ Tests/MapboxMapsTests/Viewport/Transitions/Default/DefaultViewportTransitionAnimationTests.swift
@@ -35,7 +35,7 @@ final class DefaultViewportTransitionAnimationTests: XCTestCase {
             completion(true)
         }
 
-        wait(for: [completionExpectation], timeout: 0.5)
+        wait(for: [completionExpectation], timeout: 2)
 
         XCTAssertEqual(completionStub.invocations.map(\.parameters), [true])
     }
@@ -56,7 +56,7 @@ final class DefaultViewportTransitionAnimationTests: XCTestCase {
             completion(idx != failingIndex)
         }
 
-        wait(for: [completionExpectation], timeout: 0.5)
+        wait(for: [completionExpectation], timeout: 2)
 
         XCTAssertEqual(completionStub.invocations.map(\.parameters), [false])
     }